gestionale_flask_v1.21/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
flask-migrate==4.0.5
flask-login==0.6.3
wtforms==3.0.1
email_validator==2.0.0
openpyxl==3.1.2
python-dateutil==2.8.2
Werkzeug==2.3.7
Jinja2==3.1.2